Handle empty command case
When a blank line is entered as a shell command, the time log would
print the time since the previous command was run (instead of the 0 time
it took to run nothing) and the name of the previous command. Now the
previous command name and time are unset at the end of log_command_time
and if either are empty at the beginning, then just the current
timestamp is printed.

Also committed .gitignore to prevent committing the history file by
accident.